+++ scripts/test.sh
@@ -53,6 +53,7 @@ cargo test --manifest-path=opentelemetry-sdk/Cargo.toml --all-features trace::sp
 cargo test --manifest-path=opentelemetry-sdk/Cargo.toml --all-features trace::span_processor::tests::simple_self_diagnostics_counter_records_already_shutdown_drops -- --ignored --exact
 cargo test --manifest-path=opentelemetry-sdk/Cargo.toml --all-features logs::batch_log_processor::tests::self_diagnostics_counter_records_queue_full_drops -- --ignored --exact
 cargo test --manifest-path=opentelemetry-sdk/Cargo.toml --all-features logs::batch_log_processor::tests::self_diagnostics_counter_records_already_shutdown_drops -- --ignored --exact
+cargo test --manifest-path=opentelemetry-sdk/Cargo.toml --all-features logs::batch_log_processor::tests::self_diagnostics_queue_size_and_capacity -- --ignored --exact
 cargo test --manifest-path=opentelemetry-sdk/Cargo.toml --all-features logs::simple_log_processor::tests::self_diagnostics_counter_records_success -- --ignored --exact
 cargo test --manifest-path=opentelemetry-sdk/Cargo.toml --all-features logs::simple_log_processor::tests::self_diagnostics_counter_records_already_shutdown_drops -- --ignored --exact
 cargo test --manifest-path=opentelemetry-sdk/Cargo.toml --all-features logs::logger::tests::log_created_counts_intake_without_processors -- --ignored --exact